Off-by-1 issue with slider<->step# calcs
When the Step # changed, or the SliderValue changed, in Steps mode, the calculation to convert step# to value or vice versa had some off-by-one issues.  The main visible symptom was that the slider position didn't always update when clicking the PreviousCue button (in a 6-step chaser it would update every-other-click).  Also, if moving the slider position using up/down arrows, the step # didn't change at the exact same slider positions as were jumped to using the Next-Cue button.