[change] Clone scene before adding to new track in v4 showmanager
When a Scene is added to a new Showmanager track, it becomes the base scene of a new sequence. As a result, it was getting marked as "Hidden" (though it didn't actually disappear from view until the workspace was closed and re-opened...).  For some users, losing access to the imported Scene is unwanted behavior.  Especially if the Scene was already referenced somewhere else, such as from a VCButton or Collection, because it remained referenced but couldn't ever be re-selected because it was hidden.

This change creates a copy of the Scene before creating the new Sequence, so that the cloned copy is what gets marked as hidden. This preserves the original Scene along with any other references to it and keeps it available for future use.