[change] Make v4 soloframe mirroring more like v5
In a soloframe, if a function starts due to a parent action initiated from outside the soloframe, the associated widget shows an orange border and gets assigned a tri-state value of Monitoring.

In v5, this does not affect the active widget in the soloframe, i.e. whatever the user last activated stays activated.

But in v4, this causes any other active widget in the soloframe to turn off its associated function.  The user is also prevented from turning off an active vc button in a soloframe if its function is also active due to another parent.  And, if a widget in a soloframe (i.e. a collection or chaser) Activates and causes another widget in the soloframe to change to Monitoring mode, then the just-Activated widget turns off (and so does the monitoring widget).

This change makes v4 work more like v5.  Specifically, the Monitoring mode is now just a border color indication, and does not prevent turning off any VC button nor does it cause an Active widget in the soloframe to turn off.