[change] Make v4 soloframe more like v5
In a soloframe, if a function starts due to action from outside the soloframe, the associated widget shows an orange border and gets assigned a tri-state value of Monitoring.

In v5, this does not affect the active widget in the soloframe, i.e. whatever the user last activated stays activated.

But in v4, this causes any other Active function associated with a widget in the soloframe to turn off.  The user is also prevented from turning off an active button (if its function is also active due to another parent).  And, if a widget in a soloframe (i.e. a collection or chaser) causes another widget in the soloframe to change to Monitoring mode then the Active widget turns off (and so does the monitoring widget).

This change makes v4 work more like v5.