External Input Direct modes for automation software
Added optional alternate external input modes to VCButton and VCCuelist, to allow more direct control from external automation inputs such as MultiPlay or Ableton Live.

New mode for Buttons allows external input to turn a button ON or OFF based on the value being >0 or =0, rather than Toggling between on and off whenever value is >0.  This allows an external cue that is intended to turn a button ON to be triggered multiple times without unwanted consequences.

New mode for Cuelists allows external input to activate a specific chaser step by step #, rather than by proportional positioning.  There's a "Direct Step #" mode which accepts all DMX values 1-255, and a "Direct Step # (MIDI)" mode which uses the MIDI range of 1-127 (within the Cuelist widget it divides input values by 2 before selecting the step, because MIDI inputs are multiplied by 2 earlier in the input processing chain).  In both cases, jump directly to the requested step # now, rather than calculating a fader level based on the requested step #.

Fixed a corner case in qmlui where the bottom label wasn't updating if I:  1) jump to the last step, and then 2) jump to the n-2th step (it was previously only updating the label if the "next step" changed, but in this case, the next-step calculated to the same n-1th step).